# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/webhooks":{"get":{"tags":["webhook-controller"],"summary":"Get all webhooks","description":"Returns a list of webhooks from database","operationId":"getAll","responses":{"200":{"description":"Successfully retrieved webhooks list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WebhookDTO"}}}}},"404":{"description":"Webhooks not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WebhookDTO"}}}}}}},"post":{"tags":["webhook-controller"],"summary":"Register webhook","description":"Api to register a webhook to be triggered for each payment","operationId":"registerWebhook","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WebhookDTO"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WebhookDTO"}}}},"400":{"description":"Invalid url supplied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WebhookDTO"}}}},"200":{"description":"Successfully registered the webhook","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WebhookDTO"}}}}}}},"/payments":{"get":{"tags":["payment-controller"],"summary":"Get all Payments","description":"Returns a list of payments","operationId":"getAll_1","responses":{"404":{"description":"Payments not found","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentDTO"}}}}},"200":{"description":"Successfully retrieved payments list","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentDTO"}}}}}}},"post":{"tags":["payment-controller"],"summary":"Save or update a Payment","description":"Receives a payment and save it on database.","operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentDTO"}}},"required":true},"responses":{"200":{"description":"Successfully saved the payment","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentDTO"}}}},"500":{"description":"Internal Server Error when saving the payment","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentDTO"}}}},"400":{"description":"Invalid card number supplied","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaymentDTO"}}}}}}},"/webhooks/reprocess":{"get":{"tags":["payment-webhook-controller"],"summary":"Get the failed webhooks","description":"Returns a list of webhooks that failed after a payment trigger","operationId":"getAll_2","responses":{"200":{"description":"Successfully retrieved webhooks must be reprocessed","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentWebhookDTO"}}}}},"404":{"description":"There are no failed webhooks","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PaymentWebhookDTO"}}}}}}}}},"components":{"schemas":{"WebhookDTO":{"type":"object","properties":{"url":{"type":"string"},"type":{"type":"string","enum":["PAYMENT"]},"params":{"type":"array","items":{"$ref":"#/components/schemas/WebhookHeader"}}}},"WebhookHeader":{"type":"object","properties":{"name":{"type":"string"},"value":{"type":"string"}}},"PaymentDTO":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"zipCode":{"type":"string"},"cardNumber":{"type":"string"}}},"PaymentWebhookDTO":{"type":"object","properties":{"payment":{"$ref":"#/components/schemas/PaymentDTO"},"webhook":{"$ref":"#/components/schemas/WebhookDTO"},"status":{"type":"string","enum":["FAIL","SUCCESS"]}}}}}}