# src/main/java/com/ezycollect/payment/domain/dto/PaymentDTO.java
package com.ezycollect.payment.domain.dto;

import jakarta.persistence.Column;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class PaymentDTO {
    private String firstName;
    private String lastName;
    private String zipCode;
    private String cardNumber;
}