# src/main/java/com/ezycollect/payment/domain/entity/WebhookHeader.java
package com.ezycollect.payment.domain.entity;

import com.fasterxml.jackson.annotation.JsonIgnore;
import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Entity
@Table(name = "webhook_headers")
public class WebhookHeader {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @JsonIgnore
    private Long id;

    private String name;
    private String value;

    @JsonIgnore
    @ManyToOne
    @JoinColumn(name = "webhook_id")
    private Webhook webhook;
}